requirements.txt
opencv-python
ultralytics-yolov5
numpy
pygetwindow
Pillow
websocket-client